backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.1
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
python-dateutil==2.8.2
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0 